yank tc from k_tracer
